Google Viz - Word & Phrase Tree
This code can be used to generate meaningful word and phrase tree based on customer reviews and sentiments. This code specifically explores PubG game's customer review and the color of the phrase is based on the sentiment of those phrases. To be able to see the results the user must click on the phrase or a word, example "game" to be able to see the larger context.